summarizer_service/requirements.txt
playwright-stealth
Flask
requests
openai
newspaper3k
playwright
trafilatura
lxml_html_clean
python-dotenv